requirements.txt
asttokens==2.0.5
dnspython==2.2.1
esprima==4.0.1
pyIsEmail==2.0.1
func_timeout==4.3.5
python-gitlab==2.10.1
github3.py==3.2.0
GitPython==3.1.20
networkx==2.5.1
protobuf==3.19.4
python_dateutil==2.8.2
python_magic==0.4.26
pytz==2020.1
django==4.1.1
colorama==0.4.5
rarfile==3.0
requests==2.25.0
six==1.11.0
tldextract==3.1.2
pyyaml==6.0
